feat: updateContactTable() and BigInteger
add method which helps to keep contact table up to date; get rid of BigInteger in Moneybird classes and use Long instead

+++ src/main/java/com/example/demo/models/moneybird/MoneybirdLedgerAccount.java
@@ -4,14 +4,13 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Data
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class MoneybirdLedgerAccount {
     private String id;
-    private BigInteger administrationId;
+    private Long administrationId;
     private String name;
     private String accountType;
     private String accountId;
